smart-projects/cloud4rpi/requirements.txt
paho-mqtt
pep8
pylint
mock
pyfakefs
twine
pypandoc
